Move pagination template tag away from lite_forms
