+++ unit_tests/core/builtins/test_custom_tags.py
@@ -397,3 +397,27 @@ def test_str_time_on_date(input_string, expected_output):
 )
 def test_get_parties_status_optional_documents(parties, expected_status):
     assert custom_tags.get_parties_status_optional_documents(parties) == expected_status
+
+
+@pytest.mark.parametrize(
+    "url, page, expected",
+    [
+        (
+            "https://www.example.com/?page=1&example_param=foo&example_param=bar",
+            2,
+            "https://www.example.com/?page=2&example_param=foo&example_param=bar",
+        ),
+        (
+            "https://www.example.com/?page=3&example_param=bar",
+            5,
+            "https://www.example.com/?page=5&example_param=bar",
+        ),
+        (
+            "https://www.example.com/?example_param=bar",
+            "10",
+            "https://www.example.com/?example_param=bar&page=10",
+        ),
+    ],
+)
+def test_pagination_params(url, page, expected):
+    assert custom_tags.pagination_params(url, page) == expected
