+++ unit_tests/lite_forms/templatetags/test_custom_tags.py
@@ -1,0 +1,14 @@
+import pytest
+
+from lite_forms.templatetags import custom_tags
+
+
+@pytest.mark.parametrize(
+    "filename,expected",
+    [
+        ("test.pdf", "pdf"),
+        ("test", "file"),
+    ],
+)
+def test_file_type(filename, expected):
+    assert expected == custom_tags.file_type(filename)
